NetworkCreate: add parameters (#34)
* Test new parameter documentation

Trying a new layout for parameter documentation on a couple class docs.  If this is accepted by the markdown linter, this looks like a decent way to represent parameters, and should be relatively easy to automate.

* NetworkCreate: remove enable_ir

enable_ir (enableIR) is a read-only value for networks and is inherited from the fabric in which the network resides.  Removing this property since the user should not be setting it.

* NetworkCreate: Add parameters

1. Add multiple parameters found in Default_Network_Universal template.

2. util/make_markdown.py
    - Utility to generate markdown from a NDFC template.

3. lib/ndfc_python/markdown_from_template.py
    - Class used by make_markdown.py

4. util/default_network_universal_property_map.yaml
    - parameter mapping file used by make_markdown.py

5. lib/ndfc_python/network_create.py
    - Add multiple parameters from the Default_Network_Universal template.

6. docs/classes/NetworkCreate.md
    - Document the parameters added in 5 above

* Appease linters

* Appease linters