vrf_attach: support for VPC peers
VPC Peers require both peers to be included in the lanAttachList of the payload.

This PR includes the modifications needed for this requirement.

Also, this PR removes the need to include switch serial numbers in the YAML config.  The user now only needs to include the switch name (which is translated by the library for inclusion in the payload.