NetworkCreate: Remove unneeded validation
1. lib/ndfc_python/network-create.py

Remove validation from the following properties:

- gateway_ip_address

This is already validated before being passed to NetworkCreate.  An argument could be made that removing the validation within NetworkCreate class ties the class to the example script, but we’re willing to live with this in exchange for simpler code.

- gateway_ipv6_address

The validation was incorrect, since a comma-separate list of ipv6 host addresses can be passed to this property.  Hence, we are removing the validation for now and will add it to the Pydantic validator for network_create later.